feat: enabled chunking of user files
