+++ server/controllers/upload.controller.js
@@ -1,38 +1,32 @@
 const path = require('path');
 const fs = require('fs');
-const pdfParse = require('pdf-parse');
+const extractText = require('../utils/extractText.util');
+const chunkText = require('../utils/chunkText.util');
 
 const handleFileUpload = async(req, res) => {
+    const filePath = req?.file.path;
+
     try{
         if(!req.file) return res.status(400).json({error: "No file uploaded"});
+        const extractedText = await extractText(filePath);
+        const chunks = await chunkText(extractedText);
+        const uploadDirectory = path.join(__dirname, "../../temp");
+        if(!fs.existsSync(uploadDirectory)) fs.mkdirSync(uploadDirectory);
+        const outputPath = path.join(uploadDirectory, "chunks.json")
 
-        const filePath = req.file.path;
-        const originalName = req.file.originalname;
-        const fileExtension = path.extname(originalName).toLowerCase();
-
-        let extractedText = "";
-
-        try {
-            if(fileExtension == ".pdf"){
-                const data = fs.readFileSync(filePath);
-                extractedText = await pdfParse(data).then(data => data.text);
-            }
-            else if(fileExtension == ".txt"){
-                extractedText = fs.readFileSync(filePath, 'utf8')
-            }
-            else return res.status(400).json({error: "Unsupported file type"});
-        } finally {
-            fs.unlinkSync(filePath);       
-        }
-
+        await fs.promises.writeFile(outputPath, JSON.stringify(chunks, null, 2));
         return res.status(200).json({
             message: "File uploaded successfully",
-            text: extractedText
+            chunksCount: chunks.length
         });
     }
     catch (e){
         console.error("Upload Failed:", e);
         return res.status(500).json({error: "File upload failed"});
+    } finally {
+        if(filePath && fs.existsSync(filePath)){
+            fs.unlink(filePath, (err) => {if(err) console.error("Failed to delete file:", err)});
+        }
     }
 }
